rm dist, gitignored, build updated
