+++ .gitignore
@@ -2,6 +2,8 @@
 .vscode
 .history
 
+dist/
+
 node_modules
 **/*.js
 **/*.js.map
