--- src/interface.d.ts
@@ -1,3 +1,0 @@
-export interface IHash<T> {
-    [key: string]: T;
-}
