+++ src/libs/bignumber.ts
@@ -1,9 +1,7 @@
 import BigNumber from 'bignumber.js';
 
-
 BigNumber.config({
-    ROUNDING_MODE: BigNumber.ROUND_DOWN
+    ROUNDING_MODE: BigNumber.ROUND_DOWN,
 });
-
 
 export default BigNumber;
