--- src/money/interface.d.ts
@@ -1,4 +1,0 @@
-export interface IMoneyJSON {
-    assetId: string;
-    tokens: string;
-}
