--- src/pair/interface.d.ts
@@ -1,4 +1,0 @@
-export interface IAssetPairJSON {
-    amountAsset: string;
-    priceAsset: string;
-}
